requirements.txt
flask
flask-cors
requests
python-telegram-bot
dotenv
nest_asyncio
gunicorn